refactor(api): merge tutorial-video config into /config/builder
Replaces the standalone /private/config/builder/tutorial-video endpoint
introduced in 83f07dfa2 with a nested `tutorialVideo` field on the
existing /private/config/builder response. The CLI fetches both the
Android (Google OAuth) and iOS (tutorial video) builder config in one
round-trip instead of two.

Response shape:

  { enabled: true,  clientId, clientSecret, scopes, tutorialVideo }
  { enabled: false,                                  tutorialVideo }

The two halves are independent — `enabled` (OAuth) and
`tutorialVideo.enabled` (video) gate on different secret groups, and
either can be on while the other is off. iOS users who don't need
Google OAuth still get the video block when its secrets are set.

`tutorialVideo` shape (unchanged from the deleted standalone endpoint):

  { enabled: true,  presignedUrl, sha1, youtubeFallback, expiresInSeconds }
  { enabled: false }

Reads the same seven worker secrets the previous endpoint did:

  BUILDER_TUTORIAL_VIDEO_R2_BUCKET
  BUILDER_TUTORIAL_VIDEO_R2_PATH
  BUILDER_TUTORIAL_VIDEO_R2_ACCOUNT_ID
  BUILDER_TUTORIAL_VIDEO_R2_ACCESS_KEY
  BUILDER_TUTORIAL_VIDEO_R2_SECRET_KEY
  BUILDER_TUTORIAL_VIDEO_SHA1
  BUILDER_TUTORIAL_VIDEO_YOUTUBE_URL

Backwards-compatible with the existing CLI parser
(cli/src/build/onboarding/android/oauth-config.ts) — it does
`if (!parsed.enabled) return null` on the OAuth-config path and
ignores extra fields, so adding `tutorialVideo` doesn't change Android
onboarding behaviour.

Files:
  - Removes: supabase/functions/_backend/private/builder_tutorial_video.ts
  - Removes: the extra route registration in cloudflare_workers/api/index.ts
  - Extends: supabase/functions/_backend/private/config_builder.ts with
    the buildTutorialVideoConfig helper (reads + presigns) and inlines
    the field into both OAuth-enabled/disabled response branches.
