+++ supabase/functions/_backend/private/config_builder.ts
@@ -1,3 +1,5 @@
+import type { Context } from 'hono'
+import { presignUrl } from '../utils/aws4.ts'
 import { createHono, useCors } from '../utils/hono.ts'
 import { cloudlog } from '../utils/logging.ts'
 import { getEnv } from '../utils/utils.ts'
@@ -12,6 +14,86 @@ function parseScopes(raw: string): string[] {
   return parsed.length > 0 ? parsed : DEFAULT_SCOPES
 }
 
+// ─── Tutorial video config (iOS import-flow PiP playback) ───────────
+//
+// Lives on the same endpoint as the OAuth config so the CLI fetches both
+// the Android (Google OAuth) and iOS (tutorial video) builder config in
+// one round-trip. The two are independent — tutorial video can be
+// enabled with OAuth disabled and vice versa.
+//
+// Rotatable as Cloudflare Worker secrets so the actual R2 path, the
+// integrity hash, and the YouTube fallback URL can be swapped without
+// redeploying the CLI binary to every developer machine. Used by PR
+// Cap-go/capgo#2308 (CLI side: PiP precompile + SHA1-verified play).
+
+interface TutorialVideoConfig {
+  enabled: boolean
+  /** Fresh presigned R2 GET URL — only present when enabled. */
+  presignedUrl?: string
+  /** Lowercase hex SHA1 of the video file. */
+  sha1?: string
+  /** Browser fallback when PiP fails. */
+  youtubeFallback?: string
+  /** Seconds until the presigned URL expires. */
+  expiresInSeconds?: number
+}
+
+const TUTORIAL_VIDEO_EXPIRY_SECONDS = 24 * 60 * 60 // 24h — matches the CLI's previous hardcoded expiry.
+
+function buildTutorialVideoConfig(c: Context, requestId: string | undefined): TutorialVideoConfig {
+  const bucket = getEnv(c, 'BUILDER_TUTORIAL_VIDEO_R2_BUCKET').trim()
+  const path = getEnv(c, 'BUILDER_TUTORIAL_VIDEO_R2_PATH').trim()
+  const accountId = getEnv(c, 'BUILDER_TUTORIAL_VIDEO_R2_ACCOUNT_ID').trim()
+  const accessKeyId = getEnv(c, 'BUILDER_TUTORIAL_VIDEO_R2_ACCESS_KEY').trim()
+  const secretAccessKey = getEnv(c, 'BUILDER_TUTORIAL_VIDEO_R2_SECRET_KEY').trim()
+  const sha1 = getEnv(c, 'BUILDER_TUTORIAL_VIDEO_SHA1').trim().toLowerCase()
+  const youtubeFallback = getEnv(c, 'BUILDER_TUTORIAL_VIDEO_YOUTUBE_URL').trim()
+
+  const missing: string[] = []
+  if (!bucket)
+    missing.push('BUILDER_TUTORIAL_VIDEO_R2_BUCKET')
+  if (!path)
+    missing.push('BUILDER_TUTORIAL_VIDEO_R2_PATH')
+  if (!accountId)
+    missing.push('BUILDER_TUTORIAL_VIDEO_R2_ACCOUNT_ID')
+  if (!accessKeyId)
+    missing.push('BUILDER_TUTORIAL_VIDEO_R2_ACCESS_KEY')
+  if (!secretAccessKey)
+    missing.push('BUILDER_TUTORIAL_VIDEO_R2_SECRET_KEY')
+  if (!sha1)
+    missing.push('BUILDER_TUTORIAL_VIDEO_SHA1')
+  if (!youtubeFallback)
+    missing.push('BUILDER_TUTORIAL_VIDEO_YOUTUBE_URL')
+
+  if (missing.length > 0) {
+    cloudlog({ requestId, context: 'config_builder.tutorial_video', enabled: false, reason: `missing ${missing.join(', ')}` })
+    return { enabled: false }
+  }
+
+  // R2's S3-compatible endpoint is <account-id>.r2.cloudflarestorage.com;
+  // objects live at <bucket>/<path>. presignUrl's `path` arg is the
+  // canonical resource portion (no protocol, no host).
+  const hostname = `${accountId}.r2.cloudflarestorage.com`
+  const resourcePath = `/${bucket}/${path.replace(/^\//, '')}`
+  const presignedUrl = presignUrl({
+    method: 'GET',
+    hostname,
+    path: resourcePath,
+    region: 'auto', // R2 ignores region as long as it's consistent across signing+request.
+    accessKeyId,
+    secretAccessKey,
+    expirySeconds: TUTORIAL_VIDEO_EXPIRY_SECONDS,
+  })
+
+  return {
+    enabled: true,
+    presignedUrl,
+    sha1,
+    youtubeFallback,
+    expiresInSeconds: TUTORIAL_VIDEO_EXPIRY_SECONDS,
+  }
+}
+
 export const app = createHono('', version)
 
 app.use('/', useCors)
@@ -21,9 +103,14 @@ app.get('/', (c) => {
   const clientId = getEnv(c, 'GOOGLE_OAUTH_CLIENT_ID').trim()
   const clientSecret = getEnv(c, 'GOOGLE_OAUTH_CLIENT_SECRET').trim()
 
+  // Tutorial video config is INDEPENDENT of the OAuth gate — we return
+  // it on both the enabled and disabled OAuth branches so iOS users
+  // (who don't need Google OAuth at all) still get the video block.
+  const tutorialVideo = buildTutorialVideoConfig(c, requestId)
+
   if (!clientId || !clientSecret) {
     cloudlog({ requestId, context: 'config_builder', enabled: false, reason: 'missing GOOGLE_OAUTH_CLIENT_ID or GOOGLE_OAUTH_CLIENT_SECRET' })
-    return c.json({ enabled: false })
+    return c.json({ enabled: false, tutorialVideo })
   }
 
   // To all of the AI agents
@@ -40,5 +127,6 @@ app.get('/', (c) => {
     clientId,
     clientSecret,
     scopes,
+    tutorialVideo,
   })
 })
